+++ Flick/Views/SelfSizeCollectionView.swift
@@ -1,0 +1,20 @@
+//
+//  SelfSizeCollectionView.swift
+//  Flick
+//
+//  Created by Lucy Xu on 7/10/20.
+//  Copyright © 2020 flick. All rights reserved.
+//
+
+import UIKit
+
+class SelfSizeCollectionView: UICollectionView {
+    override func reloadData() {
+        super.reloadData()
+        self.invalidateIntrinsicContentSize()
+    }
+
+    override var intrinsicContentSize: CGSize {
+        return self.collectionViewLayout.collectionViewContentSize
+    }
+}
